src/scraper/requirements.txt
beautifulsoup4==4.13.3
certifi==2025.1.31
charset-normalizer==3.4.1
greenlet==3.1.1
idna==3.10
iniconfig==2.0.0
packaging==24.2
playwright==1.50.0
pluggy==1.5.0
pyee==12.1.1
pytest==8.3.4
python-dotenv==1.0.1
requests==2.32.3
selectolax==0.3.27
soupsieve==2.6
tqdm==4.67.1
typing_extensions==4.12.2
urllib3==2.3.0
